Restructured install process

+++ firstrun.sh
@@ -18,16 +18,26 @@ if [ -z "$PLEX_VERSION" ]; then
 fi
 if [ "$PLEX_VERSION" = "$INSTALLED" ]; then
     echo "Version not changed - $PLEX_VERSION"
+    /etc/service/plex/run
 else
-    echo "Updating to $PLEX_VERSION from $INSTALLED"
     # Don't uninstall the old version of plex if the download fails
     wget -q "${PLEX_URL}" -O /tmp/plexmediaserver_${PLEX_VERSION}_amd64.deb
     if [ $? -eq 0 ]; then
-        mv /etc/default/plexmediaserver /tmp/
-        apt-get remove --purge -y plexmediaserver
+        if [ -n "$INSTALLED" ]; then
+            echo "Updating Plex to $PLEX_VERSION from $INSTALLED"
+            echo "Uninstalling  Plex version $INSTALLED"
+            apt-get remove --purge -y plexmediaserver
+        fi
+        echo "Installed to Plex version $PLEX_VERSION"
         gdebi -n /tmp/plexmediaserver_${PLEX_VERSION}_amd64.deb
-        mv /tmp/plexmediaserver /etc/default/
+        /etc/init.d/plexmediaserver stop
+        # Replace default config with our own
+        cat /default_plexmediaserver > /etc/default/plexmediaserver
+        # Fix a Debianism of plex's uid being 101
+        usermod -u 999 plex
+        usermod -g 100 plex
         rm -f /tmp/plexmediaserver_${PLEX_VERSION}_amd64.deb
+        /etc/service/plex/run
         echo $PLEX_VERSION > /tmp/version
     else
         echo "Download failed, please try again later"
